pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "captures-to-md"
version = "0.1.0"
description = "Local-first CLI daemon that watches a directory and ingests dropped files via the Extend Parse API."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "j1ngg" }]
dependencies = [
    "extend-ai>=1.9,<2",
    "watchdog>=4,<7",
    "typer>=0.12,<1",
    "rich>=13,<15",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.2,<3",
    "httpx>=0.27,<1",
    "python-dotenv>=1,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<7",
    "pytest-httpx>=0.30,<1",
    "ruff>=0.5,<1",
]

[project.scripts]
captures-to-md = "captures_to_md.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/captures_to_md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"
